Merge the Experimental (#8)
* the grinder

an attemp at adding the grinder
no it is not working yet
not at all
you can get and see the block however it has no gui nor will it write to NBT yet
some parts are ALL comment the manager may wind up not being used at this rate

* more work on the Grinder

added more stuff still not done
still 100% non working
the block makes holes in the world
still need a png made still need so much more
but progress is made each time
maybe 2 more days of it and i'll have it working
just a bit past that ready to go to master branch.

* filled in some more blanks

there are still a lot of issues
some majore issues are the lack of a Gui when you click the thing also you can't see the block idk what i did to make that happen
there are still more than a few things to do but a bit more tinkering with it and a png to use as a gui would be big steps

* did a lot did not make a list

just read the detailed log github makes i'm at wit's end here i'm useing system.err to make logs to see what gets called and not a lot gets called don't know what i'm missing now

* got it to open

holy shit did a ton

* finish up the GUI

the GUI is working for the most part
will need to make rthe meat before any more GUI related issues can be found

* Doing some clean up work to Files for merge

* Revert "Doing some clean up work to Files for merge"

This reverts commit 8fb895940d6e64b578b735a3a6071151f4adc05c.

* Merge Update

Merge is working now

* Cleaning and Item/Block adding

Cleaned up a lot of code and sorted it. Renamed a lot of files into a category system to further organize ourselves. Added Ground items and most of our blocks.

* Adding Textures and a Model

Added a bunch of textures and a model for the brewer.

* IT WORKS

did enough can merge to master branch now